Adding all tests/etc. that hadn't been added; updated gitignore.
`generate_results.sh` is used to generate all of the files in `results` that
are included in the repository. `generate_results.sh` is ran on every update
to the simulation to ensure that changes to the source code either 1. do not
modify the results files, or 2. modify the results files in
intentional/predictable ways that are explicitly checked and updated in the
repo.

License: MIT
Signed-off-by: David Grisham <dgrisham@mines.edu>